Implemented Clone() API for B+Tree
A Clone()'s return is a write-able snapshot of the specified
B+Tree at the time the call... sharing only nodes that have
not been modified by either the clone or the underlying B+Tree
from which it was cloned.

Note that any "dirty" nodes in the original B+Tree will also
be "dirty" in the clone... and the "root" node will definitely
be "dirty"... so be sure to call Flush() on the clone if you
need it preserved... even if you don't modify it. And, of course,
the original B+Tree remains as it was ("clean" or "dirty").