Added support for B+Tree UpdateCloneSource()
The idea here is that if you take a Clone, the call Flush() on the Clone,
that you may want to update source of the Clone with the particulars of
where you've Flush()'d the Clone to... for the nodes that they effectively
share in common. This is particularly the pattern if, when taking the Clone,
you passed andTouch==true... and never actually Flush() the source of
the Clone... only the Clone itself.