Added ability to reverse the effect of Flush() that marks B+Tree nodes clean
The idea here is that if the result of the Flush isn't actually recorded
persistently until after Flush() returns, it's possible that the act of
persisting this information fails. In such a case, a subsequent Flush()
may want to also flush both the newly dirty nodes as well as those nodes
previously marked clean during the prior Flush().