Remove incorrect Clone() functionality
The issue with Clone() is that it's primary use case where a
snapshot is taken and then the Clone() is what is flushed was
problematic when that Flush() call failed.

If a use case for preserving a B+Tree's state exclusive of
flushing (the clone) is envisioned, a better implementation
should be crafted.