Ensure DiscardNode() is faithfully called
Previously, there were a number of cases where DiscardNode() would
not be called. Most particularly, during DeleteBy*() rebalancing,
it is possible that two nodes are merged and/or the height of the
tree is reduced. In these cases, the node going away was not marked
as being discarded and would then never be "free'd" from the backing
store for the B+Tree.

The other big opportunity for leakage is when a B+Tree is being
disposed of via a call to Discard(). Any nodes of the B+Tree not
currently in memory would not be scheduled for calls to DiscardNode().