Convert staleOnDiskReferencesList to use an underlying map
Previously, the use of a list.List would allow a discarded B+Tree node
to be free'd twice if, say, it happened to be first marked "dirty" and
later marked "to be discarded". This change takes advantage of the fact
that structs may be used as keys to a map (if they are used "by value"
instead of "by reference") and that duplicate modifications of a key
are collapsed if the value is empty since the 2nd write overwrites the
first. Here, we are using a struct{} as a value to get Go's map to
emulate a set. Finally, while iterating over a map does not guarantee
the order of returned keys, it is sufficient for the purposes of a list
use case where all elements will be "consumed".